feat: Implement player tutorial feature
- `player_tutorial_screen.dart`: Newly implemented the UI for the player tutorial. The tutorial consists of 4 slides, and the final slide displays a completion button with an animation.
- `pubspec.yaml`: Added the `sensors_plus` package to detect device rotation on the last page of the tutorial.
