+++ frontend/pubspec.yaml
@@ -57,6 +57,7 @@ dependencies:
   ffmpeg_kit_flutter_new: ^4.0.0
   hive_generator: ^2.0.1
   flutter_staggered_grid_view: ^0.7.0
+  sensors_plus: ^7.0.0
 
 dev_dependencies:
   flutter_test:
